Integrate OptoForce and change the receive procedure
To account for different modules that receive CAN messages (BLMC and OPTO),
the procedure by which frames are received and handled has to be changed:
Instead of having a receive function in each module, the CAN frame is
directly received in the user code via CAN_receiveFrame() and is then passed
to "processing" function of the modules.  This way the frame can be shown to
different modules one after another.

Further added doxygen to optoforce_can.h