Send to the interface from which we also receive.
The get rid of the hard coded send-to-interface, simply use the same which is
also used for receiving messages.  Note that this means that a CanHandle that
is used to listen to all interfaces (by setting interface=NULL) *cannot* be
used to send frames!  In this case a seperate handle with a concrete interface
specified has to be used for sending.