Add non-blocking homing method to BlmcJointModule
Add a non-blocking method `update_homing` to `BlmcJointModule` that only
performs one iteration of the homing.  This way it can be called in a
loop for each joint to run the homing of each joint simultaneously.
This method uses a simple position profile to move until the next
encoder index is found.  This index is then used as home position.

Add a `HomingState` struct to keep the state of homing between the calls
of `update_homing` and a method `init_homing` to initialize it.

Add a method `execute_homing` to `BlmcJointModules` that runs the homing
of all joints until finished.

Add methods `set_position_control_gains` and
`execute_position_controller` to `BlmcJointModule` to avoid redundant
controller implementations in different functions and to make it
possible to use different gains for different joints.

Use the new methods for homing in the `RealFinger` class.