Initialise CAN frame structs to zero
The `can_frame` struct has some unused padding members whose values are
ignored in classic CAN.  However, with CAN FD part of this padding area
is used for some FD-specific flags (compare the `canfd_frame` struct).

When using classic CAN with a device that supports CAN FD, random values
in the padding areas can be interpreted as some FD-related flags and
thus may result in the CAN controller to switch to FD mode (which could
mess up the communication if the other devices on the bus do not support
FD).

To make sure, this does not happen, explicitly initialise the whole
struct to zero before filling it with date.

In practice this is probably only relevant in `send_frame` but to be on
the save side let's consistently initialise all occurrences of
`can_frame_t`.